+++ CHANGELOG.md
@@ -75,3 +75,4 @@ These are the changes to each version that has been released on the official Vis
 <!--Publish the extension for VS2005/2008-->
 <!--sign the extension - not required for auto updates ! activity log ? github.com/mwrock/PowerShell/tree/master/src/signing-->
 <!--allow sorting of rules-->
+<!--http://sourcebrowser.io/Browse/GregTrevellick/AutoFindReplace-->